\name{compSpec}
\alias{compSpec}
\title{Reduce column dimensions}
\usage{
  compSpec(spectra, window, colab)
}
\arguments{
  \item{spectra}{matrix where each row is a spectrum and
  each column a wavelength}

  \item{window}{window size over which the spectra will be
  averaged}

  \item{colab}{???}
}
\description{
  This function reduces the column dimensions
}
\author{
  Brendan Malone
}

