\name{goof}
\alias{goof}
\title{Goodness of fit}
\usage{
  goof(observed, predicted, ...)
}
\arguments{
  \item{observed}{numeric vector of observed values}

  \item{predicted}{numeric vector of predicted values}

  \item{...}{arguments passed to \code{\link{eqscplot}}}
}
\description{
  Goodness of fit
}
\author{
  Brendan Malone
}

