\name{wavelet_smooth}
\alias{wavelet_smooth}
\title{Wavelet smoothing}
\usage{
  wavelet_smooth(spectra, res)
}
\arguments{
  \item{spectra}{a matrix or data.frame where the columns
  are wavelengths and the rows, individual spectra}

  \item{res}{level to be extracted from wavelet
  decomposition model}
}
\description{
  Wavelet smoothing
}
\author{
  Brendan Malone
}

