\name{soil_colour_spectra}
\alias{soil_color_spectra}
\alias{soil_colour_spectra}
\title{Load spectra and extract RGB and Munsell colour}
\usage{
  soil_colour_spectra(spectra, wavelengths)
}
\arguments{
  \item{spectra}{matrix or data.frame with spectra as rows
  and wavelengths as columns}

  \item{wavelengths}{vector of wavelengths corresponding to
  the columns of spectra}
}
\description{
  Converts spectra reflectance into RGB and Munsell colours
}
\author{
  Michael Nelson
}

